requirements.txt
PyYAML==6.0.3
Pygments==2.19.2
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.74.1
anyio==4.11.0
astroid==4.0.2
beautifulsoup4==4.14.2
black==25.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
dill==0.4.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
fastapi==0.121.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
isort==7.0.0
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.1.0
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.9
langgraph==1.0.3
langsmith==0.4.46
markdown-it-py==4.0.0
mccabe==0.7.0
mdurl==0.1.2
mypy_extensions==1.1.0
numpy==2.3.5
openai==2.8.1
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pipdeptree==2.30.0
platformdirs==4.5.0
pluggy==1.6.0
pyarrow==22.0.0
pydantic-settings==2.12.0
pydantic==2.12.4
pydantic_core==2.41.5
pylint==4.0.3
pymongo==4.15.4
pytest==9.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.3.0
pytz==2025.2
regex==2025.11.3
requests-toolbelt==1.0.0
requests==2.32.5
rich==14.2.0
ruff==0.14.6
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.50.0
tenacity==9.1.2
tiktoken==0.12.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
xxhash==3.6.0
zstandard==0.25.0
